publisher/requirements.txt
Flask==2.1.2
werkzeug==2.2.3
google-cloud-pubsub==2.19.1
google-auth==2.40.0
protobuf==4.25.1
grpcio==1.73.0
